% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{AnalyticsInfo}
\alias{AnalyticsInfo}
\title{Firebase Dynamic Links API Objects 
Firebase Dynamic Links API enables third party developers to programmatically create and manage Dynamic Links.}
\usage{
AnalyticsInfo(itunesConnectAnalytics = NULL, googlePlayAnalytics = NULL)
}
\arguments{
\item{itunesConnectAnalytics}{iTunes Connect App Analytics}

\item{googlePlayAnalytics}{Google Play Campaign Measurements}
}
\value{
AnalyticsInfo object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:50:17
filename: /Users/mark/dev/R/autoGoogleAPI/googlefirebasedynamiclinksv1.auto/R/firebasedynamiclinks_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AnalyticsInfo Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tracking parameters supported by Dynamic Link.
}
