% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{AndroidInfo}
\alias{AndroidInfo}
\title{AndroidInfo Object}
\usage{
AndroidInfo(androidPackageName = NULL, androidMinPackageVersionCode = NULL,
  androidLink = NULL, androidFallbackLink = NULL)
}
\arguments{
\item{androidPackageName}{Android package name of the app}

\item{androidMinPackageVersionCode}{Minimum version code for the Android app}

\item{androidLink}{If specified, this overrides the ‘link’ parameter on Android}

\item{androidFallbackLink}{Link to open on Android if the app is not installed}
}
\value{
AndroidInfo object
}
\description{
AndroidInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Android related attributes to the Dynamic Link.
}
