% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{CreateShortDynamicLinkRequest}
\alias{CreateShortDynamicLinkRequest}
\title{CreateShortDynamicLinkRequest Object}
\usage{
CreateShortDynamicLinkRequest(dynamicLinkInfo = NULL,
  longDynamicLink = NULL, suffix = NULL)
}
\arguments{
\item{dynamicLinkInfo}{Information about the Dynamic Link to be shortened}

\item{longDynamicLink}{Full long Dynamic Link URL with desired query parameters specified}

\item{suffix}{Short Dynamic Link suffix}
}
\value{
CreateShortDynamicLinkRequest object
}
\description{
CreateShortDynamicLinkRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to create a short Dynamic Link.
}
\seealso{
Other CreateShortDynamicLinkRequest functions: \code{\link{shortLinks.create}}
}
