% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{CreateShortDynamicLinkResponse}
\alias{CreateShortDynamicLinkResponse}
\title{CreateShortDynamicLinkResponse Object}
\usage{
CreateShortDynamicLinkResponse(warning = NULL, shortLink = NULL,
  previewLink = NULL)
}
\arguments{
\item{warning}{Information about potential warnings on link creation}

\item{shortLink}{Short Dynamic Link value}

\item{previewLink}{Preivew link to show the link flow chart}
}
\value{
CreateShortDynamicLinkResponse object
}
\description{
CreateShortDynamicLinkResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to create a short Dynamic Link.
}
