% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{DynamicLinkInfo}
\alias{DynamicLinkInfo}
\title{DynamicLinkInfo Object}
\usage{
DynamicLinkInfo(androidInfo = NULL, analyticsInfo = NULL,
  dynamicLinkDomain = NULL, link = NULL, iosInfo = NULL,
  socialMetaTagInfo = NULL)
}
\arguments{
\item{androidInfo}{Android related information}

\item{analyticsInfo}{Parameters used for tracking}

\item{dynamicLinkDomain}{Dynamic Links domain that the project owns, e}

\item{link}{The link your app will open, You can specify any URL your app can handle}

\item{iosInfo}{iOS related information}

\item{socialMetaTagInfo}{Parameters for social meta tag params}
}
\value{
DynamicLinkInfo object
}
\description{
DynamicLinkInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a Dynamic Link.
}
