% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{DynamicLinkWarning}
\alias{DynamicLinkWarning}
\title{DynamicLinkWarning Object}
\usage{
DynamicLinkWarning(warningMessage = NULL, warningCode = NULL)
}
\arguments{
\item{warningMessage}{The warning message to help developers improve their requests}

\item{warningCode}{The warning code}
}
\value{
DynamicLinkWarning object
}
\description{
DynamicLinkWarning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Dynamic Links warning messages.
}
