% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{GooglePlayAnalytics}
\alias{GooglePlayAnalytics}
\title{GooglePlayAnalytics Object}
\usage{
GooglePlayAnalytics(utmSource = NULL, gclid = NULL, utmCampaign = NULL,
  utmContent = NULL, utmMedium = NULL, utmTerm = NULL)
}
\arguments{
\item{utmSource}{Campaign source; used to identify a search engine, newsletter, or other}

\item{gclid}{[AdWords autotagging parameter](https://support}

\item{utmCampaign}{Campaign name; used for keyword analysis to identify a specific product}

\item{utmContent}{Campaign content; used for A/B testing and content-targeted ads to}

\item{utmMedium}{Campaign medium; used to identify a medium such as email or cost-per-click}

\item{utmTerm}{Campaign term; used with paid search to supply the keywords for ads}
}
\value{
GooglePlayAnalytics object
}
\description{
GooglePlayAnalytics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for Google Play Campaign Measurements.[Learn more](https://developers.google.com/analytics/devguides/collection/android/v4/campaigns#campaign-params)
}
