% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{IosInfo}
\alias{IosInfo}
\title{IosInfo Object}
\usage{
IosInfo(iosIpadFallbackLink = NULL, iosIpadBundleId = NULL,
  iosCustomScheme = NULL, iosBundleId = NULL, iosFallbackLink = NULL,
  iosAppStoreId = NULL)
}
\arguments{
\item{iosIpadFallbackLink}{If specified, this overrides the ios_fallback_link value on iPads}

\item{iosIpadBundleId}{iPad bundle ID of the app}

\item{iosCustomScheme}{Custom (destination) scheme to use for iOS}

\item{iosBundleId}{iOS bundle ID of the app}

\item{iosFallbackLink}{Link to open on iOS if the app is not installed}

\item{iosAppStoreId}{iOS App Store ID}
}
\value{
IosInfo object
}
\description{
IosInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
iOS related attributes to the Dynamic Link..
}
