% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{ITunesConnectAnalytics}
\alias{ITunesConnectAnalytics}
\title{ITunesConnectAnalytics Object}
\usage{
ITunesConnectAnalytics(ct = NULL, mt = NULL, pt = NULL, at = NULL)
}
\arguments{
\item{ct}{Campaign text that developers can optionally add to any link in order to}

\item{mt}{iTune media types, including music, podcasts, audiobooks and so on}

\item{pt}{Provider token that enables analytics for Dynamic Links from within iTunes}

\item{at}{Affiliate token used to create affiliate-coded links}
}
\value{
ITunesConnectAnalytics object
}
\description{
ITunesConnectAnalytics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for iTunes Connect App Analytics.
}
