% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_functions.R
\name{shortLinks.create}
\alias{shortLinks.create}
\title{Creates a short Dynamic Link given either a valid long Dynamic Link ordetails such as Dynamic Link domain, Android and iOS app information.The created short Dynamic Link will not expire.Repeated calls with the same long Dynamic Link or Dynamic Link informationwill produce the same short Dynamic Link.The Dynamic Link domain in the request must be owned by requester'sFirebase project.}
\usage{
shortLinks.create(CreateShortDynamicLinkRequest)
}
\arguments{
\item{CreateShortDynamicLinkRequest}{The \link{CreateShortDynamicLinkRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/firebase
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/firebase)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://firebase.google.com/docs/dynamic-links/}{Google Documentation}

Other CreateShortDynamicLinkRequest functions: \code{\link{CreateShortDynamicLinkRequest}}
}
