% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{SocialMetaTagInfo}
\alias{SocialMetaTagInfo}
\title{SocialMetaTagInfo Object}
\usage{
SocialMetaTagInfo(socialDescription = NULL, socialTitle = NULL,
  socialImageLink = NULL)
}
\arguments{
\item{socialDescription}{A short description of the link}

\item{socialTitle}{Title to be displayed}

\item{socialImageLink}{An image url string}
}
\value{
SocialMetaTagInfo object
}
\description{
SocialMetaTagInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Parameters for social meta tag params.Used to set meta tag data for link previews on social sites.
}
