% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebasedynamiclinks_objects.R
\name{Suffix}
\alias{Suffix}
\title{Suffix Object}
\usage{
Suffix(option = NULL)
}
\arguments{
\item{option}{Suffix option}
}
\value{
Suffix object
}
\description{
Suffix Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Short Dynamic Link suffix.
}
